requirements.txt
Crawl4AI==0.4.247
python-dotenv==1.0.1
pydantic==2.10.6
aiohttp==3.9.3
pandas==2.1.4
groq==0.4.1
argparse==1.4.0